ProcedureCall instruction
